fix failing test cases
