+++ tests/unitary/with_extras/operator/feature-store/test_marketplace_utils.py
@@ -1,8 +1,10 @@
 import os
 from unittest.mock import Mock, patch
 
+from kubernetes.client import V1PodCondition
+
 from ads.opctl.backend.marketplace.marketplace_utils import set_kubernetes_session_token_env, get_docker_bearer_token, \
-    export_helm_chart, list_container_images
+    export_helm_chart, list_container_images, wait_for_pod_ready
 
 
 def test_set_kubernetes_session_token_env():
@@ -38,3 +40,30 @@ def test_list_container_image(oci_factory: Mock):
     listing_details = Mock()
     list_container_images(listing_details)
     oci_factory.return_value.artifacts.list_container_images.assert_called_once()
+
+@patch('ads.opctl.backend.marketplace.marketplace_utils.time')
+@patch('kubernetes.client.CoreV1Api')
+@patch('kubernetes.config.load_kube_config')
+def test_wait_for_pod_unhealthy(kube_config: Mock, kube_client: Mock, timer: Mock):
+    timer.time = Mock(return_value=0.0)
+    kube_client.return_value.list_namespaced_pod.return_value.items.__getitem__.return_value.status.conditions.__iter__ = Mock(
+        return_value=iter([V1PodCondition(type="Ready", status="False")]))
+
+    assert wait_for_pod_ready('namespace', 'pod_name') == -1
+    kube_client.return_value.list_namespaced_pod.assert_called()
+    kube_config.assert_called()
+
+@patch('ads.opctl.backend.marketplace.marketplace_utils.time')
+@patch('kubernetes.client.CoreV1Api')
+@patch('kubernetes.config.load_kube_config')
+def test_wait_for_pod_healthy(kube_config: Mock, kube_client: Mock, timer: Mock):
+    timer.time = Mock(side_effect=[0.0, 11 * 60])
+    kube_client.return_value.list_namespaced_pod.return_value.items.__getitem__.return_value.status.conditions.__iter__ = Mock(
+        return_value=iter([V1PodCondition(type="Ready", status="True")]))
+
+    assert wait_for_pod_ready('namespace', 'pod_name') == 0
+    kube_client.return_value.list_namespaced_pod.assert_called_once_with(namespace='namespace',
+                                                                         label_selector='app.kubernetes.io/instance=pod_name')
+    kube_config.assert_called()
+
+
